Add test_threading
